backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pydantic==2.9.0
pydantic-settings==2.5.0
torch==2.4.0
torchvision==0.19.0
Pillow==10.4.0
numpy==1.26.4
git+https://github.com/openai/CLIP.git
ftfy
regex
